(PC-14860)[PRO] refactor: add core/Providers
