+++ pro/src/components/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/utils/_constants.js
@@ -2,5 +2,3 @@ export const DEFAULT_PROVIDER_OPTION = {
   id: 'default',
   displayName: 'Choix de la source',
 }
-
-export const CINEMA_PROVIDER_NAMES = ['ciné office']
