+++ pro/src/components/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManagerV2/VenueProviderItemV2/VenueProviderItemV2.tsx
@@ -1,10 +1,7 @@
 import React from 'react'
 
 import { getProviderInfo } from 'components/pages/Offers/domain/getProviderInfo'
-import {
-  isAllocineProvider,
-  isCinemaProvider,
-} from 'components/pages/Offers/domain/localProvider'
+import { isAllocineProvider, isCinemaProvider } from 'core/Providers'
 import { pluralize } from 'utils/pluralize'
 import { formatLocalTimeDateString } from 'utils/timezone'
