+++ pro/src/components/pages/Offers/domain/localProvider.js
@@ -1,5 +1,5 @@
-import { CINEMA_PROVIDER_NAMES } from 'components/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/utils/_constants'
-import { SYNCHRONIZED_OFFER_EDITABLE_FIELDS } from 'components/pages/Offers/Offer/OfferDetails/_constants'
+import { SYNCHRONIZED_OFFER_EDITABLE_FIELDS } from 'core/Offers'
+import { isAllocineProvider } from 'core/Providers'
 
 const doesLastProviderExist = offer =>
   Boolean(offer !== null && offer.lastProvider)
@@ -11,12 +11,6 @@ export const isSynchronizedOffer = offer => {
 export const isOfferFromStockProvider = (offer = null) => {
   return doesLastProviderExist(offer)
 }
-
-export const isAllocineProvider = provider =>
-  provider?.name.toLowerCase() === 'allociné'
-
-export const isCinemaProvider = provider =>
-  CINEMA_PROVIDER_NAMES.includes(provider?.name.toLowerCase())
 
 export const isAllocineOffer = (offer = null) => {
   return doesLastProviderExist(offer)
