+++ pro/src/core/Offers/adapters/getOfferIndividualAdapter/__specs__/useGetndividualOffer.spec.ts
@@ -196,6 +196,7 @@ describe('useGetOfferIndividual', () => {
       visa: '',
       withdrawalDetails: '',
       withdrawalDelay: null,
+      lastProvider: null,
       lastProviderName: null,
       stocks: [
         {
