+++ pro/src/core/Providers/constants.ts
@@ -1,0 +1,1 @@
+export const CINEMA_PROVIDER_NAMES = ['ciné office']
